Patch #1401999 by Karl Kurbjun - Rockboy color, sound support, and speedups
git-svn-id: svn://svn.rockbox.org/rockbox/trunk@8324 a1c6a512-1295-4272-9138-f99709370657

+++ apps/plugins/rockboy/pcm.h
@@ -9,7 +9,7 @@ struct pcm
 {
 	int hz, len;
 	int stereo;
-	short *buf;
+	byte *buf;
 	int pos;
 };
